Small set of improvements
At verifyLoadTesting.ts import functions from loadRegistrationTesting.ts
and removes unnecessary comments. Improves the code-style of this file
around the logic of choosing the file name (remove usage of let)

At loadRegistrationTesting.ts formats the content of the file when saved
and adds the logic that allows sharing resources with verifyLoadTesting

Lastly, at signup.ts declares an interface for Options.